splitted cmip6 model
